% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{allodb}
\alias{allodb}
\title{Estimate Tree Carbon using allodb package functions}
\usage{
allodb(DBH, Genus, Species, coords, output.all = TRUE, new.eqtable = NULL)
}
\arguments{
\item{DBH}{Diameter at breast height (cm)}

\item{Genus}{First part of Species binomial}

\item{Species}{Second part of Species binomial}

\item{coords}{either a vector of coordinates of the site or a matrix of
coordinates for each tree of longitude and latitude}

\item{output.all}{if TRUE outputs all data from processing, else just outputs carbon figures}

\item{new.eqtable}{a subset or extension of the allometric equation table. Create with allodb::new_equations}
}
\value{
Above-ground Biomass in kg. If output.all = TRUE then returns columns
'DBH', 'Genus', 'Species', 'AGB_allodb_kg', 'allodb_a', 'allodb_b',
'allodb_sigma' (where AGB = a*DBH^b+e {e ~ N(0,sigma^2}).
}
\description{
Use the aalodb R package to calculate carbon
}
\examples{
coords <- c(-0.088837,51.071610)
allodb(81.887, "Pinus", "nigra", coords, output.all = FALSE)
allodb(c(76, 76), c("Pinus","Pinus"), c("nigra", "abies"), coords)

}
\references{
Gonzalez-Akre, E., Piponiot, C., Lepore, M., & Anderson-Teixeira,
K. (2020). allodb: An R package for biomass estimation at globally
distributed extratropical forest plots. Methods in Ecology and Evolution,
11(10), 1273-1280
}
