% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{broad_sap_seedling2C}
\alias{broad_sap_seedling2C}
\title{Broadleaf seedlings and saplings to carbon}
\usage{
broad_sap_seedling2C(heightincm)
}
\arguments{
\item{heightincm}{tree height in centimetres}
}
\value{
carbon in tonnes
}
\description{
todo*
}
\note{
just uses simple linear relationship to get between measures
}
\references{
Jenkins, Thomas AR, et al. "FC Woodland Carbon Code:
Carbon Assessment Protocol (v2. 0)." (2018)
}
\author{
Justin Moat. J.Moat@kew.org, Isabel Openshaw. I.Openshaw@kew.org
}
