% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bunce}
\alias{bunce}
\title{Bunce biomass equation}
\usage{
bunce(name, type = NA, dbh)
}
\arguments{
\item{name}{species name (common or binomial)}

\item{type}{'broadleaf' or 'conifer' (optional)}

\item{dbh}{diameter at breast height}
}
\value{
biomass in kg
}
\description{
Calculates dry weight based on species and dbh
}
\examples{
bunce("Oak", 24)
bunce(name = c("Oak","Oak"), dbh = c(24,20))
}
\references{
Bunce, R. G. H. "Biomass and Production of Trees in a Mixed
Deciduous Woodland: I. Girth and Height as Parameters for the Estimation of
Tree Dry Weight" (1968)
}
\author{
Isabel Openshaw. I.Openshaw@kew.org
}
