% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_doc.R
\docType{data}
\name{buncedf}
\alias{buncedf}
\title{Bunce equation coefficient data table}
\format{
A data frame with 4 columns:
\describe{
\item{X}{Common name of the species}
\item{spcode}{Abbreviation for the species}
\item{a}{Bunce coefficient a}
\item{b}{Bunce coefficient b}
}
}
\source{
Bunce, R. G. H. "Biomass and Production of Trees in a Mixed
Deciduous Woodland: I. Girth and Height as Parameters for the Estimation of
Tree Dry Weight" (1968)
}
\usage{
buncedf
}
\description{
This dataset contains coefficients for the Bunce allometric equation.
}
\keyword{datasets}
