% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allometries.R
\name{compare_methods}
\alias{compare_methods}
\title{Calculate comparison statistics for allometry methods}
\usage{
compare_methods(results_df, returnv = "AGC")
}
\arguments{
\item{results_df}{Data frame with results from allometries function}

\item{returnv}{Either "AGC" for carbon or "AGB" for biomass}
}
\value{
List with comparison statistics
}
\description{
Calculates summary statistics comparing different allometric methods
including mean differences, coefficient of variation, and method agreement metrics
}
\examples{
results <- allometries(c("Quercus","Quercus"), c("robur","robur"), c(20, 30), c(10, 15))
compare_methods(results, "AGC")
}
\author{
Isabel Openshaw. I.Openshaw@kew.org
}
