% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{crown_biomasdf}
\alias{crown_biomasdf}
\title{Crown Biomass Data}
\format{
A data frame with 6 columns:
\describe{
\item{Species.species}{Name of the species}
\item{Code}{Code for the species}
\item{b1}{First parameter related to biomass calculation}
\item{p}{Second parameter related to biomass calculation}
\item{A}{Parameter A description}
\item{b2}{Second parameter related to biomass calculation}
}
}
\source{
Jenkins, Thomas AR, et al. "FC Woodland Carbon Code: #' Carbon
Assessment Protocol (v2. 0)." (2018). (Equation 1)
}
\usage{
crown_biomasdf
}
\description{
This dataset contains crown biomass information for various tree species.
For use with the crownbiomass (Eq 6 & 7 of the FC's WCC)
}
\keyword{datasets}
