% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crownbiomass}
\alias{crownbiomass}
\title{Forestry commission crown biomass estimates}
\usage{
crownbiomass(spcode, dbh, sig_dbh = NA, conf = 0.05)
}
\arguments{
\item{spcode}{Crown biomass species code, crown_biomasdf$Code or if not
defined for species, lookup_df$short to find relating lookup_df$Crown}

\item{dbh}{diameter at breast height in centimetres}

\item{sig_dbh}{dbh sigma}

\item{conf}{one minus percentage confidence or relative error for
coefficients of the equation}
}
\value{
biomass (oven dry tonnes)
}
\description{
Function to find crown biomass (composed of branches,
stem tips and foliage) depending on species and dbh
}
\examples{
crownbiomass("CBSP", 25)
crownbiomass("CBOK", dbh = 25, sig_dbh = 10)
crownbiomass(c("CBOK","CBOK"), dbh = c(30,50), sig_dbh = 1)
}
\references{
Jenkins, Thomas AR, et al. "FC Woodland Carbon Code:
Carbon Assessment Protocol (v2. 0)." (2018). Section 5.2.2.
}
\author{
Justin Moat. J.Moat@kew.org, Isabel Openshaw. I.Openshaw@kew.org
}
