% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Carbon_conversion.R
\name{ctoco2e}
\alias{ctoco2e}
\title{Carbon to CO2 equivalent}
\usage{
ctoco2e(carbon)
}
\arguments{
\item{carbon}{carbon}
}
\value{
carbon dioxide equivalent
}
\description{
Function to convert from carbon to carbon dioxide equivalent
}
\examples{
ctoco2e(448)
ctoco2e(c(448, 450))
}
\author{
Justin Moat. J.Moat@kew.org, Isabel Openshaw. I.Openshaw@kew.org
}
