% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{CVF_df}
\alias{CVF_df}
\title{Coefficient of Variation lookup table}
\format{
A data frame with 5 columns:
\describe{
\item{method}{reference for CVF}
\item{type}{type of tree: broadleaf or conifer}
\item{biome}{biome of study}
\item{CVF}{Percentage of coefficient of variation}
\item{confience}{Confience of estimate}
}
}
\usage{
CVF_df
}
\description{
This dataset provides a lookup table for values for the coefficient of variation for converting biomass to carbon
}
\keyword{datasets}
