% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{error_division}
\alias{error_division}
\title{Analytical error progression for a quotient}
\usage{
error_division(a, sig_a, b, sig_b, returnv = "sigmasquared")
}
\arguments{
\item{a}{numerator}

\item{sig_a}{sigma for a}

\item{b}{denominator}

\item{sig_b}{sigma for b}

\item{returnv}{return value, either 'sigma' (standard deviation) or
'sigma squared' (variance)}
}
\value{
either an estimate for sigma or sigma squared of a quotient
}
\description{
Calculates sigma squared for x when x = a / b
}
\examples{
error_division(50, 10, 100, 2)
error_division(5, 0.5, 10, 0.1)
}
\references{
Taylor, J. R. (1997). An Introduction to Error Analysis: The
Study of Uncertainties in Physical Measurements (2nd ed.). University
Science Books.
}
\author{
Isabel Openshaw. I.Openshaw@kew.org
}
