% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fc_agc}
\alias{fc_agc}
\title{Calculate above ground carbon}
\usage{
fc_agc(spcode, DBH, height, method = "Matthews1", biome, returnv = "AGC")
}
\arguments{
\item{spcode}{species code}

\item{height}{in metres}

\item{biome}{tropical, Subtropical, Mediterranean,Temperate, Boreal or all}

\item{returnv}{To return either 'AGC' \link{default} or 'All'}

\item{dbh}{in centimetres}

\item{biomass2c_method}{method of converting biomass to carbon.
See biomass2c function}
}
\value{
either Above ground carbon, AGC in tonnes, or a list of tariff
number, merchantable volume (metres cubed), stem volume (metres cubed),
stem biomass (tonnes), stem carbon (tonnes), canopy carbon (tonnes) and
root carbon (tonnes)
}
\description{
Function that inputs tree species code, DBH, height and method
for converting biomass to carbon, and returns the carbon estimate
}
\references{
Jenkins, Thomas AR, et al. "FC Woodland Carbon Code:
Carbon Assessment Protocol (v2. 0)." (2018).
}
\author{
Justin Moat. J.Moat@kew.org, Isabel Openshaw I.Openshaw@kew.org
}
