% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{global_wd}
\alias{global_wd}
\title{Retrieves wood density values (g/cm^3) from the Global Wood Density Database
based on binomial name and a specified region.
If no value exists for a species in a region, it calculates a mean based on available data from broader categories (global species mean, genus in region, genus globally, family in region, family globally).}
\usage{
global_wd(binomial, region = "World")
}
\arguments{
\item{binomial}{Character vector of binomial names (genus species).}

\item{region}{A single character string specifying the region to match.
Choose from: "Africa (extratropical)", "Africa (tropical)", "Australia",
"Australia/PNG (tropical)", "Central America (tropical)", "China", "India",
"Europe", "Mexico", "Madagascar", "NorthAmerica", "Oceania",
"South America (extratropical)", "South America (tropical)",
"South-East Asia", "South-East Asia (tropical)"}
}
\value{
Numeric vector of wood density values corresponding to the inputs. Returns NA if no relevant match is found.
}
\description{
Retrieves wood density values (g/cm^3) from the Global Wood Density Database
based on binomial name and a specified region.
If no value exists for a species in a region, it calculates a mean based on available data from broader categories (global species mean, genus in region, genus globally, family in region, family globally).
}
\examples{
global_wd(binomial = c("Quercus alba", "Pinus sylvestris"), region = "Europe")
}
\references{
Zanne, A.E., et al. (2009). Global wood density database. Dryad. http://hdl.handle.net/10255/dryad.235
}
