% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_metadata.R
\name{\%||\%}
\alias{\%||\%}
\title{NULL-coalescing operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{Value to check for NULL.}

\item{y}{Default value to return if x is NULL.}
}
\value{
x if x is not NULL, otherwise y.
}
\description{
Returns the left-hand side if not NULL, otherwise returns right-hand side.
}
\keyword{internal}
