% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lookspcode}
\alias{lookspcode}
\title{Lookup species code}
\usage{
lookspcode(name, type = NA, spcode = "short")
}
\arguments{
\item{name}{name of species (common or botanical). See lookup_data.Rda}

\item{type}{either 'broadleaf' or 'conifer'}

\item{spcode}{either 'short', 'single', 'stand', 'root' or 'crown'}
}
\value{
Species code
}
\description{
Function that looks up species codes for Woodland Carbon Code
}
\examples{
lookspcode(name="Pine")
lookspcode(name='Quercus robur', spcode='single')
lookspcode(name=c('Oak', 'Quercus', 'Quercus rubra', 'Hawthorn', NA) ,
 type = c(rep("broadleaf", 4), NA), spcode='short')
}
\references{
Jenkins, Thomas AR, et al. "FC Woodland Carbon Code:
Carbon Assessment Protocol (v2. 0)." (2018).
}
\author{
Isabel Openshaw I.Openshaw@kew.org
}
