% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{method_assumptions}
\alias{method_assumptions}
\title{Allometric Method Assumptions}
\format{
A data frame with 20 rows (5 assumptions per method) and 3 variables:
\describe{
\item{method}{Character. Method identifier matching \code{method_metadata$method}}
\item{assumption_order}{Integer. Order of the assumption (1-5)}
\item{assumption}{Character. Description of the assumption}
}
}
\source{
Package authors' compilation from original method publications
}
\usage{
method_assumptions
}
\description{
A dataset containing the key assumptions for each allometric method.
This table has a one-to-many relationship with \code{method_metadata},
with multiple assumptions per method.
}
\details{
This dataset is user-accessible and can be used to understand the key
assumptions underlying each allometric method. The data is lazy-loaded
and can be accessed directly by name.
}
\examples{
# Access the data directly (lazy-loaded automatically)
method_assumptions

# Get assumptions for WCC method
method_assumptions[method_assumptions$method == "WCC", ]

# Get first assumption for each method
method_assumptions[method_assumptions$assumption_order == 1, ]

# Get all assumptions for Bunce method
method_assumptions[method_assumptions$method == "Bunce", "assumption"]

}
\seealso{
\code{\link{method_metadata}} for general method information,
\code{\link{get_method_metadata}} for a function-based API to access this data
}
\keyword{datasets}
