% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{nominal_specific_gravity}
\alias{nominal_specific_gravity}
\title{Nominal Specific Gravity Data}
\format{
A data frame with 4 columns:
\describe{
\item{Species}{Species name}
\item{Sp_code}{Species code}
\item{Allocated.Species}{Allocated species group}
\item{Nominal.Specific.Gravity..NSG.}{Nominal specific gravity values}
}
}
\source{
Jenkins, Thomas AR, et al. "FC Woodland Carbon Code: #' Carbon
Assessment Protocol (v2. 0)." (2018). (Equation 1)
}
\usage{
nominal_specific_gravity
}
\description{
This dataset contains the nominal specific gravity for various species.
}
\keyword{datasets}
