% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Uncertainty.R
\name{parameter_uncertainty}
\alias{parameter_uncertainty}
\title{Create Parameter Uncertainty Specification}
\usage{
parameter_uncertainty(value, se = NULL, relative_se = 0.025)
}
\arguments{
\item{value}{Parameter value (coefficient)}

\item{se}{Standard error of the parameter}

\item{relative_se}{Relative standard error. Used if se is NULL.}
}
\value{
A list suitable for use in mc_uncertainty inputs
}
\description{
Helper function for allometric equation parameter uncertainty.
}
\examples{
# Bunce coefficient 'a' with SE from regression
a_input <- parameter_uncertainty(1.868, se = 0.047)

# Or with relative SE (2.5 percent of value)
b_input <- parameter_uncertainty(2.268, relative_se = 0.025)

}
