% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allometries.R
\name{plot_allometry_comparison}
\alias{plot_allometry_comparison}
\title{Create comparison plots for allometry methods}
\usage{
plot_allometry_comparison(
  results_df,
  plot_type = "bar",
  returnv = "AGC",
  show_errors = TRUE,
  interactive = FALSE,
  log_scale = FALSE,
  color_scheme = "default",
  font_size = 12,
  plot_width = NULL,
  x_min = NULL,
  x_max = NULL,
  y_min = NULL,
  y_max = NULL
)
}
\arguments{
\item{results_df}{Data frame with results from allometries function}

\item{plot_type}{Type of plot: "bar", "dbh", or "height" (dbh and height are scatter plots)}

\item{returnv}{Either "AGC" for carbon or "AGB" for biomass}

\item{show_errors}{Logical, whether to show error bars}

\item{interactive}{Logical, whether to return plotly plot}
}
\value{
ggplot or plotly object
}
\description{
Creates ggplot2 or plotly plots comparing different allometric methods
}
\examples{
results <- allometries(c("Quercus","Quercus"), c("robur","robur"), c(20, 30), c(10, 15))
plot_allometry_comparison(results, "bar", "AGC", interactive = FALSE)
}
\author{
Isabel Openshaw. I.Openshaw@kew.org
}
