% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensitivity.R
\name{plot_sensitivity}
\alias{plot_sensitivity}
\title{Convenience function for plotting sensitivity analysis}
\usage{
plot_sensitivity(x, type = "comparison", ...)
}
\arguments{
\item{x}{A sensitivity_analysis object}

\item{type}{Plot type (see \code{plot.sensitivity_analysis})}

\item{...}{Additional arguments}
}
\value{
A ggplot2 object
}
\description{
Alternative function for plotting sensitivity analysis results
that can be used when S3 method dispatch doesn't work (e.g., when sourcing
files directly).
}
