% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_allometries.R
\name{print.allodb_multi_result}
\alias{print.allodb_multi_result}
\title{Print method for allodb_multi_result}
\usage{
\method{print}{allodb_multi_result}(x, ...)
}
\arguments{
\item{x}{An allodb_multi_result object}

\item{...}{Additional arguments (unused)}
}
\description{
Formatted display for allodb multi-tree results
}
