% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allometries.R
\name{print.allometries_rich_result}
\alias{print.allometries_rich_result}
\title{Print method for allometries rich result}
\usage{
\method{print}{allometries_rich_result}(x, ...)
}
\arguments{
\item{x}{An allometries_rich_result object}

\item{...}{Additional arguments (unused)}
}
\description{
Formatted display for multi-method allometry results
}
