% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compare_allometry.R
\name{print.allometry_comparison}
\alias{print.allometry_comparison}
\title{Print method for allometry comparison}
\usage{
\method{print}{allometry_comparison}(x, ...)
}
\arguments{
\item{x}{An allometry_comparison object}

\item{...}{Additional arguments (unused)}
}
\description{
Custom print method for compare_allometries output
}
