% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_metadata.R
\name{print.allometry_result}
\alias{print.allometry_result}
\title{Print method for allometry_result}
\usage{
\method{print}{allometry_result}(x, ...)
}
\arguments{
\item{x}{An allometry_result object}

\item{...}{Additional arguments (unused)}
}
\description{
Displays the result with assumptions prominently shown
}
