% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Carbon_conversion.R, R/functions.R
\name{print.biomass2c_result}
\alias{print.biomass2c_result}
\title{Print method for biomass2c result}
\usage{
\method{print}{biomass2c_result}(x, ...)

\method{print}{biomass2c_result}(x, ...)
}
\arguments{
\item{x}{A biomass2c_result object}

\item{...}{Additional arguments (unused)}
}
\description{
Formatted display for biomass to carbon conversion results

Formatted display for biomass to carbon conversion results
}
