% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_allometries.R
\name{print.biomass_multi_result}
\alias{print.biomass_multi_result}
\title{Print method for biomass_multi_result}
\usage{
\method{print}{biomass_multi_result}(x, ...)
}
\arguments{
\item{x}{A biomass_multi_result object}

\item{...}{Additional arguments (unused)}
}
\description{
Formatted display for BIOMASS multi-tree results
}
