% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_allometries.R
\name{print.bunce_multi_result}
\alias{print.bunce_multi_result}
\title{Print method for multiple Bunce results}
\usage{
\method{print}{bunce_multi_result}(x, ...)
}
\arguments{
\item{x}{A bunce_multi_result object}

\item{...}{Additional arguments (unused)}
}
\description{
Formatted display for multi-tree Bunce results
}
