% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_WCC.R
\name{print.wcc_multi_result}
\alias{print.wcc_multi_result}
\title{Print method for wcc_multi_result}
\usage{
\method{print}{wcc_multi_result}(x, ...)
}
\arguments{
\item{x}{A wcc_multi_result object}

\item{...}{Additional arguments (unused)}
}
\description{
Formatted display for WCC multi-tree results
}
