% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Uncertainty.R
\name{pro_error_carbon}
\alias{pro_error_carbon}
\title{Carbon progression of errors}
\usage{
pro_error_carbon(
  vol,
  sig_vol,
  den,
  sig_den,
  biom,
  sig_biom,
  nruns = 10000,
  returnv = NULL
)
}
\arguments{
\item{vol}{volume (m3)}

\item{sig_vol}{sigma (standard deviation) for volume}

\item{den}{density (g/cm3)}

\item{sig_den}{sigma (standard deviation) for wood density}

\item{biom}{biomass (kg)}

\item{sig_biom}{sigma (standard deviation) for biomass}

\item{nruns}{number of iteration, suggest 10,000 as min and 100,000 is a good number}

\item{returnv}{if null then mean and sd is returned else vector of
quantiles ie c(5,50,95)/100 will return 0.05, mean and 0.95 quantiles.}
}
\value{
either vector of mean and sd or vector of quantiles
}
\description{
Progression of errors through monte carlo simulation
}
\examples{
vol <- 100
volsd <- 10
den <- 0.5
densd <- 0.005
biom <- 0.5
biomsd <- 0.0025
nruns <-100000
 pro_error_carbon(vol,volsd,den,densd,biom,biomsd,nruns=100000)
 #returns quantiles at 5\%,50\% and 95\%
 pro_error_carbon(vol,volsd,den,densd,biom,biomsd,nruns=100000,returnv=c(5,50,95)/100)
}
\references{
todo** and to write at export
}
\author{
Justin Moat. J.Moat@kew.org, Isabel Openshaw. I.Openshaw@kew.org
}
