% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Uncertainty.R
\name{residual_uncertainty}
\alias{residual_uncertainty}
\title{Create Residual Uncertainty Specification}
\usage{
residual_uncertainty(rse = NULL, cv_residual = 0.2, estimate = 1)
}
\arguments{
\item{rse}{Residual standard error of the model (on original scale)}

\item{cv_residual}{Coefficient of variation of residuals. Used if rse is NULL.}

\item{estimate}{The point estimate (used with cv_residual)}
}
\value{
A list suitable for use in mc_uncertainty inputs
}
\description{
Helper function for model residual error. This represents the
inherent prediction uncertainty of the allometric equation.
}
\details{
Residual error is often the largest component of total uncertainty in
allometric estimates. It represents the natural variation in tree allometry
that the model cannot capture.

For log-transformed models, RSE should be back-transformed appropriately.
}
\examples{
# Model with RSE of 0.15 on log scale (approx 15\% CV)
residual_input <- residual_uncertainty(rse = 50, estimate = 500)

}
