% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_doc.R
\docType{data}
\name{root_biomassdf}
\alias{root_biomassdf}
\title{Root Biomass Data}
\format{
A data frame with 5 columns:
\describe{
\item{Species.species}{Name of the species}
\item{Code}{Code for the species}
\item{b1}{Parameter b1 for biomass calculation}
\item{a}{Parameter a description}
\item{b2}{Parameter b2 for biomass calculation}
}
}
\source{
Jenkins, Thomas AR, et al. "FC Woodland Carbon Code: #' Carbon
Assessment Protocol (v2. 0)." (2018). (Equation 1)
}
\usage{
root_biomassdf
}
\description{
This dataset contains root biomass parameters for different tree species.
}
\keyword{datasets}
