% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{seedlings_conifer}
\alias{seedlings_conifer}
\title{Conifer Seedlings Data}
\format{
A data frame with 3 columns:
\describe{
\item{height.cm}{Height of the seedling in cm}
\item{Carbon.kg}{Carbon content in kg}
\item{cat}{Category (conifer)}
}
}
\source{
Jenkins, Thomas AR, et al. "FC Woodland Carbon Code: #' Carbon
Assessment Protocol (v2. 0)." (2018). (Equation 1)
}
\usage{
seedlings_conifer
}
\description{
This dataset contains carbon data for conifer seedlings.
}
\keyword{datasets}
