% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{sp_lookupdf}
\alias{sp_lookupdf}
\title{Species Lookup Table}
\format{
A data frame with 18 columns:
\describe{
\item{short}{Short species code}
\item{common_name}{Common name of the species}
\item{latin_name}{Latin name of the species}
\item{type}{Type of species (broadleaf/conifer)}
\item{level}{Classification level}
\item{single}{Single or grouped species}
\item{stand}{Stand data}
\item{NSG}{Nominal specific gravity}
\item{origin}{Species origin}
\item{Crown}{Crown type code}
\item{Root}{Root type code}
\item{notes}{Additional notes}
}
}
\source{
Jenkins, Thomas AR, et al. "FC Woodland Carbon Code: #' Carbon
Assessment Protocol (v2. 0)." (2018). (Equation 1)
}
\usage{
sp_lookupdf
}
\description{
This dataset provides a lookup table for species, including common names, Latin names, and other attributes.
}
\keyword{datasets}
