% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{stemvol}
\alias{stemvol}
\title{Stem Volume Data}
\format{
A data frame with 2 columns:
\describe{
\item{dbh..cm.}{Diameter at breast height in cm}
\item{X}{Volume in cubic meters}
}
}
\source{
Jenkins, Thomas AR, et al. "FC Woodland Carbon Code: #' Carbon
Assessment Protocol (v2. 0)." (2018). (Equation 1)
}
\usage{
stemvol
}
\description{
This dataset contains stem volume data for trees based on diameter at breast height (dbh).
}
\keyword{datasets}
