% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allometries.R
\name{summary.allometry_result}
\alias{summary.allometry_result}
\title{Summary method for allometry_result}
\usage{
\method{summary}{allometry_result}(object, ...)
}
\arguments{
\item{object}{An allometry_result object}

\item{...}{Additional arguments (unused)}
}
\description{
Quick summary of the result
}
