% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{summary_per_area}
\alias{summary_per_area}
\title{Calculate Total per unit area with Error for Biomass or Carbon}
\usage{
summary_per_area(
  input,
  sigma_input,
  area,
  sigma_area,
  returnv = "sigma",
  plots = FALSE
)
}
\arguments{
\item{input}{Vector of tree-level above-ground biomass or carbon or a list of
multiple areas to be summarised separately.}

\item{sigma_input}{Vector of standard deviations associated with input or a
list of multiple areas to be summarised separately.}

\item{area}{Total area sampled (whole habitat or sampled plots) or a vector
of multiple areas.}

\item{sigma_area}{Standard deviation of plot area measurement or a vector.}

\item{returnv}{error returned as standard deviation, 'sigma' (default) or
variance, 'sigmasquared'}

\item{plots}{if plots = TRUE then the list is treated as individual plots to
be summed as individual plots then taken the mean to get the weighted per
area mean. Default is plots = FALSE, then items in the list are treated as
separate habitats and metrics outputted seperately.}
}
\value{
A list containing the estimated total per unit area and its
propagated standard deviation.
}
\description{
This function calculates the total carbon or biomass per unit
area with propagated error.
}
\examples{
AGB <- c(2.3, 1.8, 3.2)  # Biomass estimates for trees
SD_AGB <- c(0.2, 0.15, 0.3)  # Standard deviations of biomass estimates
summary_per_area(AGB, SD_AGB, 0.5, 0.05)

AGB_2 <- c(5.9, 7.5, 2.1)  # Biomass estimates for treesz
SD_AGB_2 <- c(0.7, 1.02, 0.3)  # Standard deviations of biomass estimates

summary_per_area(input = list(AGB, AGB_2), sigma_input = list(SD_AGB,
SD_AGB_2), area = c(0.5, 0.7), sigma_area = c(0.05, 0.09))

summary_per_area(AGB, SD_AGB, 0.5, 0.09)

Carbon <- AGB * 0.5  # Convert biomass to carbon
SD_Carbon <- SD_AGB * 0.5  # Approximate error scaling
summary_per_area(Carbon, SD_Carbon, 2.5, 0.09)

}
\references{
Taylor, J. R. (1997). An Introduction to Error Analysis: The
Study of Uncertainties in Physical Measurements (2nd ed.). University
Science Books.
}
\author{
Isabel Openshaw. I.Openshaw@kew.org, Justin Moat. J.Moat@kew.org
}
