% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_doc.R
\docType{data}
\name{tarif2heightdf}
\alias{tarif2heightdf}
\title{Tarif to Height Conversion Data}
\format{
A data frame with 5 columns:
\describe{
\item{common.name}{Common name of the species}
\item{abbreviation}{Abbreviation for the species}
\item{a1}{Conversion parameter a1}
\item{a2}{Conversion parameter a2}
\item{a3}{Conversion parameter a3}
}
}
\source{
Jenkins, Thomas AR, et al. "FC Woodland Carbon Code: #' Carbon
Assessment Protocol (v2. 0)." (2018). (Equation 1)
}
\usage{
tarif2heightdf
}
\description{
This dataset contains conversion parameters from tariff values to tree height
for various species.
}
\keyword{datasets}
