% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{tariff_broaddf}
\alias{tariff_broaddf}
\title{Broadleaf Tariff Data}
\format{
A data frame with 6 columns:
\describe{
\item{Common.name}{Common name of the species}
\item{abbreviation}{Abbreviation for the species}
\item{a1}{Tariff parameter a1}
\item{a2}{Tariff parameter a2}
\item{a3}{Tariff parameter a3}
\item{a4}{Tariff parameter a4}
}
}
\source{
Jenkins, Thomas AR, et al. "FC Woodland Carbon Code: #' Carbon
Assessment Protocol (v2. 0)." (2018). (Equation 1)
}
\usage{
tariff_broaddf
}
\description{
This dataset contains tariff parameters for broadleaf species.
}
\keyword{datasets}
