% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tariff_vol_area}
\alias{tariff_vol_area}
\title{Tariff number from volume and basal area}
\usage{
tariff_vol_area(vol, dbh)
}
\arguments{
\item{vol}{tree volume in metres cubed}

\item{dbh}{diameter at breast height in centimetres}
}
\value{
Tariff number
}
\description{
Using the sample tree’s basal area and volume to calculate the
tariff number. Basal area is calculated by ba = (pi * dbh^2)/40000.
}
\references{
Jenkins, Thomas AR, et al. "FC Woodland Carbon Code:
Carbon Assessment Protocol (v2. 0)." (2018). (Equation 1)
}
\author{
Justin Moat. J.Moat@kew.org, Isabel Openshaw. I.Openshaw@kew.org
}
