% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{treevol}
\alias{treevol}
\title{Forestry commission tree wood volume}
\usage{
treevol(mtreevol, DBH, sigma_mtreevol = NA)
}
\arguments{
\item{mtreevol}{merchantable tree volume}

\item{DBH}{diameter at breast height in centimeters (greater than 6.5 cm)}

\item{sigma_mtreevol}{sigma for mtreevol (optional)}
}
\value{
volume metres cubed or if sigma_mtreevol is provided then additionally
returns the error as a list
}
\description{
Calculate the stem volume by multiplying the merchantable
tree volume by the appropriate species multiplication factor from stemvol.rda
}
\references{
Jenkins, Thomas AR, et al. "FC Woodland Carbon Code:
Carbon Assessment Protocol (v2. 0)." (2018).
}
\author{
Justin Moat. J.Moat@kew.org, Isabel Openshaw. I.Openshaw@kew.org
}
