% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_WCC.R
\name{wcc_stand_carbon}
\alias{wcc_stand_carbon}
\title{Stand-level carbon (WCC Method D) - alias}
\usage{
wcc_stand_carbon(...)
}
\arguments{
\item{...}{Arguments passed to \code{fc_stand_carbon}}
}
\value{
Same as \code{fc_stand_carbon}
}
\description{
Alias for \code{fc_stand_carbon}. Calculate stand-level above-ground
carbon using WCC Method D (top height and mean DBH from sample trees).
}
\seealso{
\code{\link{fc_stand_carbon}}
}
