% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{wd_zanne}
\alias{wd_zanne}
\title{Wood Density Dataset from Zanne et al.}
\format{
A data frame with 16,468 rows and 9 variables:
\describe{
\item{Family}{Character. Plant family.}
\item{Genus}{Character. Plant genus.}
\item{Species}{Character. Plant species.}
\item{Binomial}{Character. Genus + species name.}
\item{Wood.density}{Numeric. Wood density (g/cm³), ranging from 0.08 to 1.39.}
\item{Reference}{Integer. Reference ID for source of data.}
\item{Continent}{Character. Continent where data was collected.}
\item{Biome}{Character. Biome classification.}
\item{Region}{Character. Combined continent and biome description.}
}
}
\source{
Zanne, A.E., et al. (2009). Global wood density database. Dryad. http://hdl.handle.net/10255/dryad.235
}
\usage{
wd_zanne
}
\description{
A dataset of wood density measurements compiled by Zanne et al. across various species, continents, and biomes.
}
\keyword{datasets}
