% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WCC_functions.R
\name{woodbiomass}
\alias{woodbiomass}
\title{Forestry Commission wood biomass (WCC component)}
\usage{
woodbiomass(treevol, nsg, sig_treevol = NULL, sig_nsg = 0.09413391)
}
\arguments{
\item{treevol}{tree volume in metres cubed}

\item{nsg}{Nominal Specific Gravity (g/cm^3 or t/m^3)}

\item{sig_treevol}{tree volume sigma (optional)}

\item{sig_nsg}{sigma for NSG (default = 0.094, from UK timber data)}
}
\value{
biomass in oven dry tonnes, or if sig_treevol is provided,
a list with woodbiomass and sigma.
}
\description{
Multiply the mean total tree volume by the nominal specific
gravity (NSG) to give the stem biomass, in oven dry tonnes. This is a
component calculation within the Woodland Carbon Code methodology.
}
\examples{
woodbiomass(10, 0.56, 5)
}
\references{
Jenkins, Thomas AR, et al. "FC Woodland Carbon Code:
Carbon Assessment Protocol (v2. 0)." (2018). Lavers, G.M. and Moore, G.L.
(1983) The strength properties of timber. Building Research Establishment
Report CI/Sfb i(J3). Building Research Establishment, Garston.
}
\author{
Justin Moat. J.Moat@kew.org, Isabel Openshaw. I.Openshaw@kew.org
}
