% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{allodb}
\alias{allodb}
\title{Estimate Tree Carbon using Biomass package functions}
\usage{
allodb(DBH, Genus, Species, coords, output.all = TRUE, new.eqtable = NULL)
}
\arguments{
\item{DBH}{Diameter at breast height (cm)}

\item{Genus}{First part of Species binomial}

\item{Species}{Second part of Species binomial}

\item{coords}{either a vector of coordinates of the site or a matrix of
coordinates for each tree of longitude and latitude}

\item{output.all}{if TRUE outputs all data from processing, else just outputs carbon figures}

\item{new.eqtable}{a subset or extension of the allometric equation table. Create with allodb::new_equations}
}
\value{
your dataframe back with added columns of carbon estimates. If
output.all = FALSE, then returns columns 'Genus_corrected','Species',
'Family','Latitude','Longitude','DBH','AGB_Biomass_kg'. If output.all = TRUE
then additionally returns columns 'Wood_Density', 'Wood_Density_sd',
'Height_est', 'RSE' (Residual Standard Error of the model), 'Height_1' (which
is inputed height filled in with Height estimate where missing)
}
\description{
Using the Biomass package to calculate carbon
}
