% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allometries.R, R/functions.R
\name{allometries}
\alias{allometries}
\title{Calculate above ground carbon}
\usage{
allometries(
  genus,
  species,
  dbh,
  height,
  type = NULL,
  method = "IPCC2",
  returnv = "AGC",
  region = "Europe",
  biome = "temperate",
  coords = c(-0.088837, 51.07161),
  re_dbh = 0.05,
  re_h = 0.1,
  re = 0.025,
  nsg = NULL,
  sig_nsg = 0.09413391,
  checkTaxo = FALSE
)

allometries(
  genus,
  species,
  dbh,
  height,
  type = NULL,
  method = "IPCC2",
  returnv = "AGC",
  region = "Europe",
  biome = "temperate",
  coords = c(-0.088837, 51.07161),
  re_dbh = 0.05,
  re_h = 0.1,
  re = 0.025,
  nsg = NULL,
  sig_nsg = 0.09413391,
  checkTaxo = FALSE
)
}
\arguments{
\item{genus}{First part of species binomial}

\item{species}{Second part of species binomial}

\item{dbh}{diameter at breast height in centimetres}

\item{height}{in metres}

\item{type}{either 'broadleaf' or 'conifer'}

\item{method}{method of converting biomass to carbon. Either 'Thomas' or
'IPCC2' has the error associated and "Matthews1", "Matthews2" or "IPCC1" do not.}

\item{returnv}{either 'AGC' or 'AGB' for above ground carbon or biomass,
respectively}

\item{region}{of the World. See ?getWoodDensity for the list of regions}

\item{biome}{temperate, boreal, mediterranean, tropical, subtropical or all}

\item{coords}{either a vector of coordinates of the site or a matrix of
coordinates for each tree of longitude and latitude}

\item{re_dbh}{relative measurement error for diameter at breast height,
single value}

\item{re_h}{relative error of height measurement, single value}

\item{re}{relative error of coefficients (default = 2.5\%)}

\item{nsg}{nominal specific gravity. Optionally specified, else will use that
given by the WCC}

\item{sig_nsg}{sigma for nominal specific gravity (NSG) or wood density}

\item{checkTaxo}{If TRUE then BIOMASS::correctTaxo will check spelling of
species name. This is included in output if there were modifications made.}
}
\value{
either Above ground carbon, AGC in tonnes, or if output.all = TRUE,
a list of tariff number, merchantable volume (metres cubed), stem volume
(metres cubed), stem biomass (tonnes), stem carbon (tonnes), canopy carbon
(tonnes) and root carbon (tonnes)

a dataframe with the inputs (with duplicated columns specifing
if these were updated) and Above ground carbon, AGC in tonnes or above-ground
biomass, AGB in tonnes for each allometric method
}
\description{
Function that inputs tree species code, dbh, height and method
for converting biomass to carbon, and returns the carbon estimate

Function that inputs tree species code, dbh, height and method
for converting biomass to carbon, and returns the carbon estimate
}
\examples{
allometries("Quercus", "robur", 20, 10)
allometries("Quercus", "robur", 20, 10)
}
\references{
Jenkins, Thomas AR, et al. "FC Woodland Carbon Code:
Carbon Assessment Protocol (v2. 0)." (2018).
Réjou-Méchain, M., Tanguy, A., Piponiot, C., Chave, J., & Hérault, B. (2017).
BIOMASS: an R package for estimating above-ground biomass and its
uncertainty in tropical forests. Methods in Ecology and Evolution, 8(9),
1163-1167
Gonzalez-Akre, E., Piponiot, C., Lepore, M., & Anderson-Teixeira, K. (2020).
allodb: An R package for biomass estimation at globally distributed
extratropical forest plots. Methods in Ecology and Evolution, 11(10),
1273-1280
Bunce, R. G. H. "Biomass and Production of Trees in a Mixed Deciduous
Woodland: I. Girth and height as Parameters for the Estimation of Tree Dry
Weight" (1968)

Jenkins, Thomas AR, et al. "FC Woodland Carbon Code:
Carbon Assessment Protocol (v2. 0)." (2018).
Réjou-Méchain, M., Tanguy, A., Piponiot, C., Chave, J., & Hérault, B. (2017).
BIOMASS: an R package for estimating above-ground biomass and its
uncertainty in tropical forests. Methods in Ecology and Evolution, 8(9),
1163-1167
Gonzalez-Akre, E., Piponiot, C., Lepore, M., & Anderson-Teixeira, K. (2020).
allodb: An R package for biomass estimation at globally distributed
extratropical forest plots. Methods in Ecology and Evolution, 11(10),
1273-1280
Bunce, R. G. H. "Biomass and Production of Trees in a Mixed Deciduous
Woodland: I. Girth and height as Parameters for the Estimation of Tree Dry
Weight" (1968)
}
\author{
Justin Moat. J.Moat@kew.org, Isabel Openshaw I.Openshaw@kew.org
}
