% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{biomass}
\alias{biomass}
\title{Estimate Tree Carbon using Biomass package functions}
\usage{
biomass(
  DBH,
  Height = NULL,
  Genus,
  Species,
  coords,
  region = "World",
  output.all = TRUE
)
}
\arguments{
\item{DBH}{Diameter at breast height}

\item{Height}{Height of tree (optional, if not specified will estimate height)}

\item{Genus}{First part of Species binomial}

\item{Species}{Second part of Species binomial}

\item{coords}{either a vector of coordinates of the site or a matrix of
coordinates for each tree of longitude and latitude}

\item{region}{of the World. See ?getWoodDensity for the list of regions}

\item{output.all}{if TRUE outputs all data from processing, else just outputs carbon figures}
}
\value{
Above-ground Biomass in kg. If output.all = FALSE, then returns
columns 'Genus_corrected','Species', 'Family', 'Latitude', 'Longitude',
'DBH', 'AGB_Biomass_kg'. If output.all = TRUE then additionally returns
columns 'Wood_Density', 'Wood_Density_sd', 'Height_est', 'RSE'
(Residual Standard Error of the model), 'Height_1' (which is inputed height
filled in with Height estimate where missing).
}
\description{
Using the Biomass package to calculate carbon
}
\examples{
coords <- c(-0.088837, 51.071610)
biomass(12, 12, 'Quercus', 'robus', coords)
}
\references{
Réjou-Méchain, M., Tanguy, A., Piponiot, C., Chave, J., & Hérault
, B. (2017). BIOMASS: an R package for estimating above-ground biomass and
its uncertainty in tropical forests. Methods in Ecology and Evolution, 8(9),
1163-1167
}
