% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{BIOMASS}
\alias{BIOMASS}
\alias{biomass}
\title{Estimate Tree Carbon using Biomass package functions}
\usage{
BIOMASS(
  dbh,
  height = NULL,
  genus,
  species,
  coords,
  region = "World",
  output.all = TRUE
)
}
\arguments{
\item{dbh}{Diameter at breast height (todo units)}

\item{height}{height of tree (optional, if not specified will estimate height)}

\item{genus}{First part of species binomial}

\item{species}{Second part of species binomial}

\item{coords}{either a vector of coordinates of the site or a matrix of
coordinates for each tree of longitude and latitude}

\item{region}{of the World. See ?getWoodDensity for the list of regions}

\item{output.all}{if TRUE outputs all data from processing, else just outputs carbon figures}
}
\value{
Above-ground Biomass in kg. If output.all = FALSE, then returns
columns 'genus_corrected','species', 'Family', 'Latitude', 'Longitude',
'dbh', 'AGB_Biomass_kg'. If output.all = TRUE then additionally returns
columns 'Wood_Density', 'Wood_Density_sd', 'height_est', 'RSE'
(Residual Standard Error of the model), 'height_1' (which is inputed height
filled in with height estimate where missing).
}
\description{
Using the Biomass package to calculate carbon
}
\examples{
coords <- c(-0.088837, 51.071610)
BIOMASS(12, 12, 'Quercus', 'robus', coords)
}
\references{
Réjou-Méchain, M., Tanguy, A., Piponiot, C., Chave, J., & Hérault
, B. (2017). BIOMASS: an R package for estimating above-ground biomass and
its uncertainty in tropical forests. Methods in Ecology and Evolution, 8(9),
1163-1167
}
