% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{biomass}
\alias{biomass}
\title{Estimate Tree Carbon using Biomass package functions}
\usage{
biomass(
  DBH,
  Height = NULL,
  Genus,
  Species,
  coords,
  region = "World",
  output.all = TRUE
)
}
\arguments{
\item{DBH}{Diameter at breast height}

\item{Height}{Height of tree (optional, if not specified will estimate height)}

\item{Genus}{First part of Species binomial}

\item{Species}{Second part of Species binomial}

\item{coords}{either a vector of coordinates of the site or a matrix of
coordinates for each tree of longitude and latitude}

\item{region}{of the World. See ?getWoodDensity for the list of regions}

\item{output.all}{if TRUE outputs all data from processing, else just outputs carbon figures}
}
\value{
your dataframe back with added columns of carbon estimates. If
output.all = FALSE, then returns columns 'Genus_corrected','Species_corrected',
'Family','Latitude','Longitude','DBH','AGB_Biomass_kg'. If output.all = TRUE
then additionally returns columns 'Wood_Density', 'Wood_Density_sd',
'Height_est', 'RSE' (Residual Standard Error of the model), 'Height_1' (which
is inputed height filled in with Height estimate where missing)
}
\description{
Using the Biomass package to calculate carbon
}
