% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{biomass2c}
\alias{biomass2c}
\title{Calculates biomass conversion to carbon}
\usage{
biomass2c(biomass, method, type = NULL, biome = NULL, sigma_biomass = NA)
}
\arguments{
\item{biomass}{(usually kg or metric tonnes)}

\item{method}{Method of calculating the carbon volatile fraction.
Matthews1 - Simplest with the carbon volatile fraction, CVF = 50\% (Matthews, 1993)
Matthews2 - CVF based on type (broadleaf or conifer)
IPCC1 - CVF = 47.7\% (IPCC, 2006)
IPCC2 - Lookup CVF by type and biome
Thomas1 - CVF = 48.3\% and 95\% CI of 0.3\% (Thomas and Martin, 2012)
Thomas2 - Lookup by type and biome}

\item{type}{broadleaf or conifer. Only required for method = 'Matthews2',
'IPCC2' or 'Thomas'}

\item{biome}{tropical, subtropical, mediterranean, temperate or boreal.
Only needed for 'IPCC2' and 'Thomas' methods.}

\item{sigma_biomass}{biomass uncertainty (optional)}
}
\value{
either carbon or if 'sigma_biomass' is provided, returns a list of carbon value and error. Only associated errors with method = 'IPCC2' and 'Thomas'
}
\description{
Converts biomass to carbon values using the carbon volatile
fraction from chosen method/citation.
}
\references{
(1) Thomas, Sean C., and Adam R. Martin. "Carbon content of tree
tissues: a synthesis." Forests 3.2 (2012): 332-352.
https://www.mdpi.com/1999-4907/3/2/332.
(2) IPCC. Forest lands. Intergovernmental Panel on Climate Change Guidelines
for National Greenhouse Gas Inventories; Institute for Global Environmental
Strategies (IGES): Hayama,Japan, 2006; Volume 4, p. 83.
(3) Matthews, G.A.R. (1993) The Carbon Content of Trees. Forestry Commission
Technical Paper 4. Forestry Commission, Edinburgh. 21pp. ISBN: 0-85538-317-8
@examples
biomass2c(4, method="IPCC2", c("conifer"), "temperate")
biomass2c(c(3,4), method="IPCC2", c("conifer","conifer"), "temperate", c(0.7,1))
@export
}
\author{
Justin Moat. J.Moat@kew.org, Isabel Openshaw. I.Openshaw@kew.org
}
