% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{biomass2c}
\alias{biomass2c}
\title{Calculates biomass conversion to carbon}
\usage{
biomass2c(biomass, method, type, biome, return = "carbon")
}
\arguments{
\item{biomass}{(usually kg or metric tonnes)}

\item{type}{broadleaf or conifer. Only required for method = 'Matthews2',
'IPCC2' or 'Thomas'}

\item{biome}{tropical, subtropical, mediterranean, temperate or boreal.
Only needed for 'IPCC2' and 'Thomas' methods.}

\item{return}{either 'carbon' = just carbon value or 'error' = list of carbon
value with error. Only associated errors with method = 'IPCC2' and 'Thomas'}
}
\value{
either carbon value or list of carbon value with error
}
\description{
Converts biomass to carbon values using the carbon volatile
fraction from chosen method/citation.
}
\references{
(1) Thomas, Sean C., and Adam R. Martin. "Carbon content of tree
tissues: a synthesis." Forests 3.2 (2012): 332-352.
https://www.mdpi.com/1999-4907/3/2/332.
(2) IPCC. Forest lands. Intergovernmental Panel on Climate Change Guidelines
for National Greenhouse Gas Inventories; Institute for Global Environmental
Strategies (IGES): Hayama,Japan, 2006; Volume 4, p. 83.
(3) Matthews, G.A.R. (1993) The Carbon Content of Trees. Forestry Commission
Technical Paper 4. Forestry Commission, Edinburgh. 21pp. ISBN: 0-85538-317-8
}
\author{
Justin Moat. J.Moat@kew.org, Isabel Openshaw. I.Openshaw@kew.org
}
