% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{broadleaf_tariff}
\alias{broadleaf_tariff}
\title{Carbon tariff number for broadleaf tree}
\usage{
broadleaf_tariff(spcode, height, dbh, sig_dbh = NA, sig_h = NA, conf_a = 0.05)
}
\arguments{
\item{spcode}{species code (single)}

\item{height}{tree height in meters}

\item{dbh}{diameter at breast height in centimetres}

\item{sig_dbh}{sigma for diameter at breast height (optional)}

\item{sig_h}{sigma for tree height (optional)}

\item{conf_a}{1 - the percentage confidence interval for coefficients or
the relative error of coefficients (default = 5\%)}
}
\value{
tariff number and error if sigma of variables inputted
}
\description{
Use dbh and tree height to derive the tariff number of each
sample tree. Species-specific estimates of a1 – a4 are found in the
R data file, 'tariff_broaddf'.
}
\examples{
broadleaf_tariff(spcode = 'OK', height = 25, dbh = 75)
broadleaf_tariff(spcode = "OK", height = 25, dbh = 15, sig_dbh = 5, sig_h = 1)
broadleaf_tariff(spcode = "OK", height = 24, dbh = 15, sig_dbh = 5, sig_h = 1)
}
\references{
Jenkins, Thomas AR, et al. "FC Woodland Carbon Code:
Carbon Assessment Protocol (v2. 0)." (2018). Method B, Equation 2.
}
\author{
Justin Moat. J.Moat@kew.org, Isabel Openshaw. I.Openshaw@kew.org
}
