% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{broadleaf_tariff}
\alias{broadleaf_tariff}
\title{Carbon tariff number for broadleaf tree}
\usage{
broadleaf_tariff(spcode, height, dbh, sigma_dbh = NA, sigma_height = NA)
}
\arguments{
\item{spcode}{species code (single)}

\item{height}{tree height in meters}

\item{dbh}{diameter at breast height in centimetres}

\item{sigma_dbh}{sigma for diameter at breast height (optional)}

\item{sigma_height}{sigma for tree height (optional)}
}
\value{
tariff number and error if sigma of variables inputted
}
\description{
Use DBH and tree height to derive the tariff number of each
sample tree. Species-specific estimates of a1 – a4 are found in the
R data file, 'tariff_broaddf'.
}
\examples{
broadleaf_tariff(spcode = 'OK', height = 25, dbh = 1.5)
55.96704

}
\references{
Jenkins, Thomas AR, et al. "FC Woodland Carbon Code:
Carbon Assessment Protocol (v2. 0)." (2018). Method B, Equation 2.
}
\author{
Justin Moat. J.Moat@kew.org, Isabel Openshaw. I.Openshaw@kew.org
}
