% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WCC_functions.R
\name{broadleaf_tariff}
\alias{broadleaf_tariff}
\title{Carbon tariff number for broadleaf tree}
\usage{
broadleaf_tariff(spcode, height, dbh, re_dbh = NULL, re_h = 0.1, re = 0.025)
}
\arguments{
\item{spcode}{species code (single)}

\item{height}{tree height in metres}

\item{dbh}{diameter at breast height in centimetres}

\item{re_dbh}{relative error for dbh/percentage measurement error (optional)}

\item{re_h}{relative error of height measurement (optional)}

\item{re}{relative error of coefficients (default = 0.025)}
}
\value{
tariff number or if relative errors are provided returns a list of
tariff number and an estimate for sigma
}
\description{
Use dbh and tree height to derive the tariff number of each
sample tree. species-specific estimates of a1-a4 are found in the
R data file, 'tariff_broaddf'.
}
\examples{
broadleaf_tariff(spcode = 'OK', height = 25, dbh = 75)
broadleaf_tariff(spcode = "OK", height = 25, dbh = 15, re_dbh = 0.05, re_h = 0.1)
broadleaf_tariff(spcode = "OK", height = 24, dbh = 15, re_dbh = 0.05, re_h = 0.1)
}
\references{
Jenkins, Thomas AR, et al. "FC Woodland Carbon Code:
Carbon Assessment Protocol (v2. 0)." (2018). Method B, Equation 2.
}
\author{
Justin Moat. J.Moat@kew.org, Isabel Openshaw. I.Openshaw@kew.org
}
