% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bunce}
\alias{bunce}
\title{Bunce biomass equation}
\usage{
bunce(spcode, DBH)
}
\arguments{
\item{spcode}{species code}

\item{DBH}{diameter at breast height}
}
\value{
biomass in kg
}
\description{
Calculates dry weight based on species and DBH
}
\references{
Bunce, R. G. H. "Biomass and Production of Trees in a Mixed
Deciduous Woodland: I. Girth and Height as Parameters for the Estimation of
Tree Dry Weight" (1968)
}
\author{
Isabel Openshaw. I.Openshaw@kew.org
}
