% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bunce.R, R/functions.R
\docType{data}
\name{bunce}
\alias{bunce}
\title{buncedf Data Frame
Lookup table for covariables for the Bunce equation}
\format{
A data frame with 4 variables:
\describe{
\item{X}{Species common name}
\item{spcode}{Species code}
\item{a}{Covariable, a, for use within the Bunce equation}
\item{b}{Covariable, b, for use within the Bunce equation}
}
}
\source{
Bunce, R. G. H. "Biomass and Production of Trees in a Mixed
Deciduous Woodland: I. Girth and Height as Parameters for the Estimation of
Tree Dry Weight" (1968)
}
\usage{
bunce

bunce(spcode, dbh)
}
\arguments{
\item{spcode}{species code}

\item{dbh}{diameter at breast height}
}
\value{
biomass in kg
}
\description{
Calculates dry weight based on species and DBH
}
\references{
Bunce, R. G. H. "Biomass and Production of Trees in a Mixed
Deciduous Woodland: I. Girth and Height as Parameters for the Estimation of
Tree Dry Weight" (1968)
}
\author{
Isabel Openshaw. I.Openshaw@kew.org
}
\keyword{datasets}
