% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bunce}
\alias{bunce}
\title{Bunce biomass equation}
\usage{
bunce(name, dbh, type = NA, re_dbh = NULL, re = 0.025)
}
\arguments{
\item{name}{species name (common or binomial)}

\item{dbh}{diameter at breast height}

\item{type}{'broadleaf' or 'conifer' (optional)}

\item{re_dbh}{relative measurement error for diameter at breast height (optional)}

\item{re}{relative error of coefficients (default = 2.5\%)}
}
\value{
biomass in kg
}
\description{
Calculates dry weight based on species and dbh
}
\examples{
bunce("Oak", 24)
bunce(c("Oak", "Beech"), c(24,23))
}
\references{
Bunce, R. G. H. "Biomass and Production of Trees in a Mixed
Deciduous Woodland: I. Girth and Height as Parameters for the Estimation of
Tree Dry Weight" (1968)
}
\author{
Isabel Openshaw. I.Openshaw@kew.org
}
