% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{c2co2e}
\alias{c2co2e}
\title{Carbon to CO2 equivalent}
\usage{
c2co2e(carbon)
}
\arguments{
\item{carbon}{carbon}
}
\value{
carbon dioxide equivalent
}
\description{
Function to convert from carbon to carbon dioxide equivalent
}
\examples{
c2co2e(448)
c2co2e(c(448, 450))

}
\author{
Justin Moat. J.Moat@kew.org, Isabel Openshaw. I.Openshaw@kew.org
}
