% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{conifer_tariff}
\alias{conifer_tariff}
\title{Conifer tree tariff number}
\usage{
conifer_tariff(spcode, height, dbh, sigma_h = NA, sigma_dbh = NA)
}
\arguments{
\item{spcode}{species code (single)}

\item{height}{tree height in metres}

\item{dbh}{diameter at breast height in centimetres}

\item{sigma_h}{uncertainty in height}

\item{sigma_dbh}{uncertainty in dbh}
}
\value{
tariff number
}
\description{
Use dbh and tree height to calculate the tariff number of each
sample tree. Species-specific estimates of a1 – a3 are found in the
R data file, 'tariff_coniferdf'.
}
\examples{
conifer_tariff("SP", 74, 24)
conifer_tariff("SP", 74, 24, 5, 5)
}
\references{
Jenkins, Thomas AR, et al. "FC Woodland Carbon Code:
Carbon Assessment Protocol (v2. 0)." (2018).
}
\author{
Justin Moat. J.Moat@kew.org, Isabel Openshaw. I.Openshaw@kew.org
}
