% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FC_Vol_Tariff.R
\name{conifer_tariff}
\alias{conifer_tariff}
\title{Conifer tree tariff number
sample tree. species-specific estimates of a1-a3 are found in the
R data file, 'tariff_coniferdf'.}
\usage{
conifer_tariff(spcode, height, dbh, re_h = NULL, re_dbh = 0.05, re = 0.025)
}
\arguments{
\item{spcode}{species code (single)}

\item{height}{tree height in metres}

\item{dbh}{diameter at breast height in centimetres}

\item{re_h}{relative error of height measurement (optional)}

\item{re_dbh}{relative error for diameter at breast height (optional)}

\item{re}{relative error of coefficients (default = 2.5\%)}
}
\value{
tariff number or if relative errors are provided returns a list of
tariff number and an estimate for sigma
}
\description{
Conifer tree tariff number
sample tree. species-specific estimates of a1-a3 are found in the
R data file, 'tariff_coniferdf'.
}
\examples{
conifer_tariff("SP", 74, 24)
conifer_tariff("SP", 74, 24, 1, 1)
}
\references{
Jenkins, Thomas AR, et al. "FC Woodland Carbon Code:
Carbon Assessment Protocol (v2. 0)." (2018).
}
\author{
Justin Moat. J.Moat@kew.org, Isabel Openshaw. I.Openshaw@kew.org
}
