% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FC_Vol_Tariff.R
\name{crownbiomass}
\alias{crownbiomass}
\title{Forestry commission crown biomass estimates}
\usage{
crownbiomass(spcode, dbh, re_d = NULL, re = 0.025)
}
\arguments{
\item{spcode}{Crown biomass species code, crown_biomasdf$Code or if not
defined for species, lookup_df$short to find relating lookup_df$Crown}

\item{dbh}{diameter at breast height in centimetres}

\item{re_d}{relative error for diameter at breast height measurement (default = 5\%)}

\item{re}{relative error of coefficients (default = 2.5\%)}
}
\value{
biomass (oven dry tonnes) and estimated sigma
}
\description{
Function to find crown biomass (composed of branches,
stem tips and foliage) depending on species and dbh
}
\examples{
crownbiomass("CBSP", 25)
crownbiomass("CBOK", dbh = 25, re_d = 0.01)
crownbiomass(c("CBOK","CBOK"), dbh = c(30,50), re_d = 0.05)
}
\references{
Jenkins, Thomas AR, et al. "FC Woodland Carbon Code:
Carbon Assessment Protocol (v2. 0)." (2018). Section 5.2.2.
}
\author{
Justin Moat. J.Moat@kew.org, Isabel Openshaw. I.Openshaw@kew.org
}
