% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crownbiomass}
\alias{crownbiomass}
\title{Forestry commission crown biomass estimates}
\usage{
crownbiomass(spcode, dbh, sigma_dbh = NA)
}
\arguments{
\item{spcode}{Crown biomass species code, crown_biomasdf$Code which
relates to sp_lookupdf$Crown}

\item{dbh}{diameter at breast height in centimetres}

\item{sigma_dbh}{dbh sigma}
}
\value{
biomass (oven dry tonnes)
}
\description{
Function to find crown biomass (composed of branches,
stem tips and foliage) depending on species and dbh
}
\examples{
crownbiomass("CBOK", 70)
crownbiomass("CBOK", 70, sigma_dbh = 10)
crownbiomass(c("CBOK","CBOK"), dbh = c(70,70), sigma_dbh = c(10,50))
}
\references{
Jenkins, Thomas AR, et al. "FC Woodland Carbon Code:
Carbon Assessment Protocol (v2. 0)." (2018). Section 5.2.2.
}
\author{
Justin Moat. J.Moat@kew.org, Isabel Openshaw. I.Openshaw@kew.org
}
