% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{error_product}
\alias{error_product}
\title{Analytical error progression for product}
\usage{
error_product(a, a_sd, b, b_sd, c = NA, c_sd = NA)
}
\arguments{
\item{a}{first variable in product}

\item{a_sd}{sigma for a}

\item{b}{second variable in product}

\item{b_sd}{sigma for b}

\item{c}{(optional) third variable in product}

\item{c_sd}{(optional) sigma for c}
}
\value{
error for x = (abs(a * b) * sqrt((a_sd / a)^2 + (b_sd / b)^2))^2
or (abs(a * b * c) * sqrt((a_sd / a)^2 + (b_sd / b)^2 + (c_sd / c)^2))^2
}
\description{
Calculates the error for x when x = a * b or x = a * b * c
}
\references{
*todo
}
\author{
Isabel Openshaw. I.Openshaw@kew.org
}
