% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{error_product}
\alias{error_product}
\title{Analytical error progression for a product or a quotient}
\usage{
error_product(
  a,
  sig_a,
  b,
  sig_b,
  c = NULL,
  sig_c = NULL,
  returnv = "sigmasquared",
  fn = NULL
)
}
\arguments{
\item{a}{first variable in function}

\item{sig_a}{sigma for a}

\item{b}{second variable in function}

\item{sig_b}{sigma for b}

\item{c}{(optional) third variable in function}

\item{sig_c}{(optional) sigma for c}

\item{returnv}{return value, either 'sigma' (standard deviation) or
'sigma squared' (variance)}

\item{fn}{function describing how the variables are related. If not specified
assumes that the variables are related by a product.}
}
\value{
either an estimate for sigma or sigma squared error propagation for a
product or a quotient of two or three variables.
}
\description{
Calculates sigma squared for f when either f = a * b, f = a / b
or f = a * b * c, f = a / b * c, f = a * b / c
}
\examples{
a = 5 ; sig_a = 0.01
b = 10 ; sig_b = 0.1
error_product(a, sig_a, b, sig_b)
c = 12 ; sig_c = 0.05
error_product(a, sig_a, b, sig_b, c, sig_c)
# error for a quotient
error_product(a, sig_a, b, sig_b, fn = a/b)
}
\references{
Taylor, J. R. (1997). An Introduction to Error Analysis: The
Study of Uncertainties in Physical Measurements (2nd ed.). University
Science Books.
}
\author{
Isabel Openshaw. I.Openshaw@kew.org, Justin Moat. J.Moat@kew.org
}
