% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{error_product}
\alias{error_product}
\title{Analytical error progression for a product or a quotient}
\usage{
error_product(
  a,
  sig_a,
  b,
  sig_b,
  c = NULL,
  sig_c = NULL,
  returnv = "sigmasquared",
  fn = NA
)
}
\arguments{
\item{a}{first variable in function}

\item{sig_a}{sigma for a}

\item{b}{second variable in function}

\item{sig_b}{sigma for b}

\item{c}{(optional) third variable in function}

\item{sig_c}{(optional) sigma for c}

\item{returnv}{return value, either 'sigma' (standard deviation) or
'sigma squared' (variance)}

\item{fn}{function describing how the variables are related. Default is fn =
NA which assumes that the variables are a product.}
}
\value{
either an estimate for sigma or sigma squared error propagation for a
product or a quotient of two or three variables.
}
\description{
Calculates sigma squared for f when either f = a * b, f = a / b
or f = a * b * c, f = a / b * c, f = a * b / c
}
\examples{
error_product(5, 0.01, 10, 0.1)
error_product(5, 0.01, 10, 0.1, 5, 0.01)
}
\references{
Taylor, J. R. (1997). An Introduction to Error Analysis: The
Study of Uncertainties in Physical Measurements (2nd ed.). University
Science Books.
}
\author{
Isabel Openshaw. I.Openshaw@kew.org, Justin Moat. J.Moat@kew.org
}
