% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{error_product}
\alias{error_product}
\title{Analytical error progression for a product}
\usage{
error_product(
  a,
  sig_a,
  b,
  sig_b,
  c = NULL,
  sig_c = NULL,
  returnv = "sigmasquared"
)
}
\arguments{
\item{a}{first variable in product}

\item{sig_a}{sigma for a}

\item{b}{second variable in product}

\item{sig_b}{sigma for b}

\item{c}{(optional) third variable in product}

\item{sig_c}{sigma for c (if c provided then provide sig_c also)}

\item{returnv}{return value, either 'sigma' (standard deviation) or
'sigma squared' (variance)}
}
\value{
either an estimate for sigma or sigma squared of a product
}
\description{
Calculates sigma squared for x when x = a * b or x = a * b * c
}
\examples{
error_product(5, 0.01, 10, 0.1)
error_product(5, 0.01, 10, 0.1, 5, 0.01)
}
\references{
Taylor, J. R. (1997). An Introduction to Error Analysis: The
Study of Uncertainties in Physical Measurements (2nd ed.). University
Science Books.
}
\author{
Isabel Openshaw. I.Openshaw@kew.org
}
