% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fc_agc}
\alias{fc_agc}
\title{Calculate above ground carbon}
\usage{
fc_agc(
  name,
  dbh,
  height,
  type = NA,
  method = "IPCC2",
  biome = "temperate",
  output.all = TRUE,
  nsg = NA
)
}
\arguments{
\item{name}{species name, either binomial or common}

\item{dbh}{diameter at breast height in centimetres}

\item{height}{in metres}

\item{type}{either 'broadleaf' or 'conifer'}

\item{method}{method of converting biomass to carbon. Either 'Thomas' or 'IPCC2' as these specify the error associated with the carbon volatile fraction}

\item{biome}{temperate, boreal, mediterranean, tropical, subtropical or all}

\item{output.all}{if TRUE (default) outputs all data from processing, else outputs carbon estimate}

\item{nsg}{nominal specific gravity. Optionally specified, else will use that
given by the WCC}
}
\value{
either Above ground carbon (AGC) in tonnes, or a list with tariff
number, merchantable volume (metres cubed), stem volume (metres cubed),
stem biomass (tonnes), stem carbon (tonnes), canopy carbon (tonnes),
root carbon (tonnes) and AGC.
}
\description{
Function that inputs tree species code, dbh, height and method
for converting biomass to carbon, and returns the carbon estimate
}
\examples{
fc_agc(spcode='OK', dbh=74, height=24, output.all = FALSE)
# Input wood density and sd from BIOMASS package
wd <- BIOMASS::getWoodDensity('Quercus', 'robur', region='Europe')
fc_agc('OK', 72, 24, nsg = wd$meanWD)
}
\references{
Jenkins, Thomas AR, et al. "FC Woodland Carbon Code:
Carbon Assessment Protocol (v2. 0)." (2018).
}
\author{
Justin Moat. J.Moat@kew.org, Isabel Openshaw I.Openshaw@kew.org
}
