% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fc_agc_error}
\alias{fc_agc_error}
\title{Calculate above ground carbon}
\usage{
fc_agc_error(
  spcode,
  dbh,
  height,
  method = "IPCC2",
  biome = "temperate",
  returnv = "All",
  sigma_dbh = 20,
  sigma_h = 8
)
}
\arguments{
\item{spcode}{species code (single)}

\item{dbh}{diameter at breast height in centimetres}

\item{height}{in metres}

\item{method}{method of converting biomass to carbon. Either 'Thomas' or 'IPCC2' as these specify the error associated with the carbon volatile fraction}

\item{biome}{temperate, boreal, mediterranean, tropical, subtropical or all}

\item{returnv}{To return either 'AGC' (default) or 'All'}

\item{sigma_dbh}{error in sampling of dbh, single value}

\item{sigma_h}{error in sampling of height, single value}
}
\value{
either Above ground carbon, AGC in tonnes, or a list of tariff
number, merchantable volume (metres cubed), stem volume (metres cubed),
stem biomass (tonnes), stem carbon (tonnes), canopy carbon (tonnes) and
root carbon (tonnes)
}
\description{
Function that inputs tree species code, dbh, height and method
for converting biomass to carbon, and returns the carbon estimate
}
\examples{
fc_agc_error(spcode='OK', dbh=74, height=24, method="IPCC2", biome="temperate", returnv ="All", sigma_dbh=10, sigma_h=1)
fc_agc_error(spcode='OK', dbh=74, height=24, method="IPCC2", biome="temperate", returnv ="AGC", sigma_dbh=10, sigma_h=1)
}
\references{
Jenkins, Thomas AR, et al. "FC Woodland Carbon Code:
Carbon Assessment Protocol (v2. 0)." (2018).
}
\author{
Justin Moat. J.Moat@kew.org, Isabel Openshaw I.Openshaw@kew.org
}
