% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fc_agc_error}
\alias{fc_agc_error}
\title{Calculate above ground carbon}
\usage{
fc_agc_error(
  spcode,
  DBH,
  height,
  method = "Matthews1",
  biome,
  returnv = "All",
  sigma_DBH,
  sigma_H
)
}
\arguments{
\item{spcode}{species code (single)}

\item{DBH}{diameter at breast height in centimetres}

\item{height}{in metres}

\item{method}{method of converting biomass to carbon. Either 'Thomas' or 'IPCC2' as these specify the error associated with the carbon volatile fraction}

\item{biome}{tropical, Subtropical, Mediterranean,Temperate, Boreal or all}

\item{returnv}{To return either 'AGC' (default) or 'All'}
}
\value{
either Above ground carbon, AGC in tonnes, or a list of tariff
number, merchantable volume (metres cubed), stem volume (metres cubed),
stem biomass (tonnes), stem carbon (tonnes), canopy carbon (tonnes) and
root carbon (tonnes)
}
\description{
Function that inputs tree species code, DBH, height and method
for converting biomass to carbon, and returns the carbon estimate
}
\examples{
fc_agc_error(spcode='OK', DBH=74, height=24, method="IPCC2", biome="temperate", returnv ="All", sigma_DBH=10, sigma_H=1)
fc_agc_error(spcode='OK', DBH=74, height=24, method="IPCC2", biome="temperate", returnv ="AGC", sigma_DBH=10, sigma_H=1)


}
\references{
Jenkins, Thomas AR, et al. "FC Woodland Carbon Code:
Carbon Assessment Protocol (v2. 0)." (2018).
}
\author{
Justin Moat. J.Moat@kew.org, Isabel Openshaw I.Openshaw@kew.org
}
