% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lookspcode}
\alias{lookspcode}
\title{Lookup species code}
\usage{
lookspcode(name, type, name_type = "botanical", returnv = "all")
}
\arguments{
\item{name}{name of species}

\item{type}{either 'broadleaf' or 'conifer'}

\item{name_type}{either 'botanical' or 'common'}

\item{returnv}{either 'all', short', 'single', 'stand', 'root' from lookupdf.Rda}
}
\value{
Species code
}
\description{
Function that looks up species codes for Woodland Carbon Code
}
\examples{
lookspcode('Quercus robur', type = "broadleaf", returnv='short')
lookspcode(c("Scots pine", "Oak"), c("conifer", "broadleaf"),name_type= "common", "short")
lookspcode(c("Scots pine", "Oak"), c("conifer", "broadleaf"))
lookspcode(c('Quercus robur', 'Quercus') , type = c("broadleaf", "broadleaf"), returnv='short')
}
\references{
Jenkins, Thomas AR, et al. "FC Woodland Carbon Code:
Carbon Assessment Protocol (v2. 0)." (2018).
}
\author{
Isabel Openshaw I.Openshaw@kew.org
}
