% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lookspcode}
\alias{lookspcode}
\title{Lookup species code}
\usage{
lookspcode(name, name_type = "botanical", classification, returnv = "all")
}
\arguments{
\item{name}{name of species}

\item{name_type}{either 'botanical' or 'common'}

\item{classification}{either 'broadleaf' or 'conifer'}

\item{returnv}{either 'all', short', 'single', 'stand', 'root' from sp_lookupdf.Rda}
}
\value{
Species code
}
\description{
Function that looks up species codes for Woodland Carbon Code
}
\examples{
lookspcode('Quercus robur', classification = 'broadleaf', returnv='short')
lookspcode(c('Quercus robur', 'Picea abies') , classification = c('broadleaf', 'conifer'), returnv='short')

}
\references{
Jenkins, Thomas AR, et al. "FC Woodland Carbon Code:
Carbon Assessment Protocol (v2. 0)." (2018).
}
\author{
Isabel Openshaw I.Openshaw@kew.org
}
