% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lookupcode}
\alias{lookupcode}
\title{Lookup species code}
\usage{
lookupcode(name, type = NULL, code = "short", returnv = "all")
}
\arguments{
\item{name}{name of species (common or botanical). See lookup_df.Rda}

\item{type}{either 'broadleaf' or 'conifer'}

\item{code}{either 'short', 'single', 'stand', 'Root' or 'Crown' Or other
column names from lookup_df.Rda}

\item{returnv}{either 'code' for just the code output or 'all' (default) with
spname and matchtype for checking}
}
\value{
species code
}
\description{
Function that looks up species codes for Woodland Carbon Code
}
\examples{
lookupcode(name="Pine")
lookupcode(name='Quercus robur', code='single')
lookupcode(name=c('Oak', 'Quercus', 'Quercus rubra', 'Hawthorn', NA) ,
 type = c(rep("broadleaf", 4), NA), code='short')
}
\references{
Jenkins, Thomas AR, et al. "FC Woodland Carbon Code:
Carbon Assessment Protocol (v2. 0)." (2018).
}
\author{
Isabel Openshaw I.Openshaw@kew.org
}
