% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{merchtreevol}
\alias{merchtreevol}
\title{Forestry merchantable volume}
\usage{
merchtreevol(dbh, tariff, sigma_dbh = NA, sigma_tariff = NA)
}
\arguments{
\item{dbh}{diameter at breast height in centimetres}

\item{tariff}{tree or stand tariff number}

\item{sigma_dbh}{diameter at breast height sigma (optional)}

\item{sigma_tariff}{tariff sigma (optional)}
}
\value{
volume metres cubed and error if sigma of variables inputted
}
\description{
Use the tree tariff number and DBH to estimate the mean
merchantable tree volume.
}
\references{
Jenkins, Thomas AR, et al. "FC Woodland Carbon Code:
Carbon Assessment Protocol (v2. 0)." (2018).
}
\author{
Justin Moat. J.Moat@kew.org, Isabel Openshaw. I.Openshaw@kew.org
}
