% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WCC_functions.R
\name{merchtreevol}
\alias{merchtreevol}
\title{Forestry merchantable volume}
\usage{
merchtreevol(dbh, tariff, re_dbh = 0.05, sig_tariff = NULL, re = 0.025)
}
\arguments{
\item{dbh}{diameter at breast height in centimetres}

\item{tariff}{tree or stand tariff number}

\item{re_dbh}{relative error for diameter at breast height (optional)}

\item{sig_tariff}{tariff sigma (optional)}

\item{re}{relative error of coefficients (default = 0.025)}
}
\value{
volume in metres cubed and error if sig of variables inputted
}
\description{
Use the tree tariff number and dbh to estimate the mean
merchantable tree volume.
}
\examples{
merchtreevol(dbh = 24, tariff = 24)
merchtreevol(dbh = 24, tariff = 24, re_dbh = 0.05, sig_tariff = 1)
}
\references{
Jenkins, Thomas AR, et al. "FC Woodland Carbon Code:
Carbon Assessment Protocol (v2. 0)." (2018).
}
\author{
Justin Moat. J.Moat@kew.org, Isabel Openshaw. I.Openshaw@kew.org
}
