% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Uncertainty.R
\name{plot.mc_uncertainty}
\alias{plot.mc_uncertainty}
\title{Plot Monte Carlo Uncertainty Results}
\usage{
\method{plot}{mc_uncertainty}(x, type = "histogram", ...)
}
\arguments{
\item{x}{An mc_uncertainty object}

\item{type}{Plot type: "histogram" (default), "density", or "both"}

\item{...}{Additional arguments passed to plotting functions}
}
\description{
Plot Monte Carlo Uncertainty Results
}
