% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rootbiomass}
\alias{rootbiomass}
\title{Forestry commission root biomass estimates}
\usage{
rootbiomass(spcode, dbh, sigma_dbh = NA)
}
\arguments{
\item{spcode}{species code}

\item{dbh}{diameter at breast height (1.3 m) in centimetres}

\item{sigma_dbh}{sigma for DBH (optional)}
}
\value{
biomass (oven dry tonnes)
}
\description{
Function to calculate the root biomass depending on species and dbh
}
\examples{
rootbiomass(spcode = 'RBRAR', dbh = 50)
rootbiomass(spcode = 'RBRAR', dbh = 50, sigma_dbh = 10)
rootbiomass(spcode = c('RBRAR','RBRAR'), dbh = c(50, 70))
}
\references{
Jenkins, Thomas AR, et al. "FC Woodland Carbon Code:
Carbon Assessment Protocol (v2. 0)." (2018). Section 5.2.3.
}
\author{
Justin Moat. J.Moat@kew.org, Isabel Openshaw. I.Openshaw@kew.org
}
