% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rootbiomass}
\alias{rootbiomass}
\title{Forestry commission root biomass estimates}
\usage{
rootbiomass(spcode, DBH, sigma_DBH = NA)
}
\arguments{
\item{spcode}{species code}

\item{DBH}{diameter at breast height (1.3 m) in centimetres}
}
\value{
biomass (oven dry tonnes)
}
\description{
Function to calculate the root biomass depending on species and DBH
}
\examples{
rootbiomass(spcode = 'RBRAR', DBH = 50)
rootbiomass(spcode = 'RBRAR', DBH = 50, sigma_DBH = 10)
rootbiomass(spcode = c('RBRAR','RBRAR'), DBH = c(50, 70))

}
\references{
Jenkins, Thomas AR, et al. "FC Woodland Carbon Code:
Carbon Assessment Protocol (v2. 0)." (2018). Section 5.2.3.
}
\author{
Justin Moat. J.Moat@kew.org, Isabel Openshaw. I.Openshaw@kew.org
}
