% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WCC_functions.R
\name{rootbiomass}
\alias{rootbiomass}
\title{Forestry commission root biomass estimates}
\usage{
rootbiomass(spcode, dbh, re_dbh = NULL, re = 0.025)
}
\arguments{
\item{spcode}{species code}

\item{dbh}{diameter at breast height (1.3 m) in centimetres}

\item{re_dbh}{relative error for diameter at breast height (optional)}

\item{re}{relative error of coefficients of the equation (default = 2.5\%)}
}
\value{
biomass (oven dry tonnes)
}
\description{
Function to calculate the root biomass depending on species and dbh
}
\examples{
rootbiomass(spcode = 'RBRAR', dbh = 50)
rootbiomass(spcode = 'RBRAR', dbh = 50, re_dbh = 0.10)
rootbiomass(spcode = c('RBRAR','RBRAR'), dbh = c(50, 70))
}
\references{
Jenkins, Thomas AR, et al. "FC Woodland Carbon Code:
Carbon Assessment Protocol (v2. 0)." (2018). Section 5.2.3.
}
\author{
Justin Moat. J.Moat@kew.org, Isabel Openshaw. I.Openshaw@kew.org
}
