% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_WCC.R
\name{sap_seedling2C}
\alias{sap_seedling2C}
\title{Seedlings and saplings to carbon}
\usage{
sap_seedling2C(height, type, re_h = NULL, re = 0.025)
}
\arguments{
\item{height}{tree height in metres}

\item{type}{'conifer' or 'broadleaf'}

\item{re_h}{relative error of height measurement (optional)}

\item{re}{relative error of estimates (default = 0.025)}
}
\value{
carbon in tonnes or if re_h provided then additionally the error
}
\description{
Calculate the total carbon content of tree seedlings
(below and above ground)
}
\note{
just uses simple linear relationship to get between measures
}
\examples{
sap_seedling2C(0.5, 'conifer')
sap_seedling2C(height = 9, type = 'broadleaf', re_h = 0.05)
}
\references{
Jenkins, Thomas AR, et al. "FC Woodland Carbon Code:
Carbon Assessment Protocol (v2. 0)." (2018)
}
\author{
Justin Moat. J.Moat@kew.org, Isabel Openshaw. I.Openshaw@kew.org
}
