% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sap_seedling2C}
\alias{sap_seedling2C}
\title{Seedlings and saplings to carbon}
\usage{
sap_seedling2C(heightincm, type, sig_height = NA, conf = 0.05)

sap_seedling2C(heightincm, type, sig_height = NA, conf = 0.05)
}
\arguments{
\item{heightincm}{tree height in centimetres}

\item{type}{'conifer' or 'broadleaf'}

\item{conf}{relative error or 1-percentage confidence interval on estimates}

\item{sig_h}{estimated measurement error of height}
}
\value{
carbon in tonnes

carbon in tonnes
}
\description{
Calculate the total carbon content of tree seedlings
(below and above ground)

Calculate the total carbon content of tree seedlings
(below and above ground)
}
\note{
just uses simple linear relationship to get between measures
}
\examples{
con_sap_seedling2C(10)
con_sap_seedling2C(50)
sap_seedling2C(50)
sap_seedling2C(heightincm = 900, type = 'broadleaf', sig_height = 5)
}
\references{
Jenkins, Thomas AR, et al. "FC Woodland Carbon Code:
Carbon Assessment Protocol (v2. 0)." (2018) Table 6.1.2.

Jenkins, Thomas AR, et al. "FC Woodland Carbon Code:
Carbon Assessment Protocol (v2. 0)." (2018)
}
\author{
Justin Moat. J.Moat@kew.org, Isabel Openshaw. I.Openshaw@kew.org
}
