% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_conversion.R
\name{sd_area}
\alias{sd_area}
\title{Standard Deviation for Measurement of the total Area}
\usage{
sd_area(perimeter, RMSE, sum_plots = FALSE)
}
\arguments{
\item{perimeter}{A numeric vector of perimeters (in meters).}

\item{RMSE}{A numeric value representing the Root Mean Square Error (RMSE)
of the perimeter measurement (in meters).}

\item{sum_plots}{if equal to TRUE then sum all of the standard deviations of
the plots. If false then outputs the standard deviation of each input.}
}
\value{
A numeric vector representing the standard deviation of the area for
each corresponding perimeter (in hectares).
}
\description{
This function calculates the standard deviation of the area
based on the perimeter and Root Mean Square Error (RMSE).
}
\examples{
# Example with multiple perimeters of 200m, 300m, and 400m.
# If measured to the nearest 50cm within 100m plot, the RMSE = 0.5m
sd_area(c(200, 300, 400), 0.5)
}
\references{
Taylor, J. R. (1997). An Introduction to Error Analysis: The
Study of Uncertainties in Physical Measurements (2nd ed.). University
Science Books.

Goodchild, M. F., & Hunter, G. J. (1997). "A simple positional
accuracy measure for linear features." International Journal of Geographical
Information Science, 11(3), 299-306.
}
\author{
Isabel Openshaw. I.Openshaw@kew.org, Justin Moat. J.Moat@kew.org
}
