% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stand_tariff}
\alias{stand_tariff}
\title{Tariff number by stand height}
\usage{
stand_tariff(spcode, height, sig_h = NA, conf_a = 0.05)
}
\arguments{
\item{spcode}{species code}

\item{height}{tree height in metres}

\item{sig_h}{sigma for height (optional)}

\item{conf_a}{1 - the percentage confidence interval for coefficients or
the relative error of coefficients (default = 5\%)}
}
\value{
either tariff number or if sig_h is provided, then returns a list
of the tariff number and uncertainty
}
\description{
Use the estimated stand top height to calculate the stand
tariff number.
}
\examples{
stand_tariff("OK", height = 10)
stand_tariff("OK", height = 10, sig_h = 1)
stand_tariff(spcode = "AH", height = 10, sig_h = 0.5)
conifer_tariff(spcode=c("SP", "NS", "OK"), c(74, 25,2), c(74, 25,2), 5, 5)
}
\references{
Jenkins, Thomas AR, et al. "FC Woodland Carbon Code:
Carbon Assessment Protocol (v2. 0)." (2018).
}
\author{
Justin Moat. J.Moat@kew.org, Isabel Openshaw. I.Openshaw@kew.org
}
