% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stand_tariff}
\alias{stand_tariff}
\title{Tariff number by stand height}
\usage{
stand_tariff(spcode, height, re_h = NULL, re = 0.025)
}
\arguments{
\item{spcode}{species code}

\item{height}{tree height in metres}

\item{re_h}{relative error of height measurement (optional)}

\item{re}{relative error of coefficients (default = 2.5\%)}
}
\value{
either tariff number or if re_h is provided, then returns a list
of the tariff number and an estimate of sigma for tariff number
}
\description{
Use the estimated stand top height to calculate the stand
tariff number based on species-specific coefficients.
}
\examples{
stand_tariff("OK", height = 10)
stand_tariff("OK", height = 10, re_h = 0.01)
stand_tariff(spcode = "AH", height = 10, re_h = 0.05)
}
\references{
Jenkins, Thomas AR, et al. "FC Woodland Carbon Code:
Carbon Assessment Protocol (v2. 0)." (2018).
}
\author{
Justin Moat. J.Moat@kew.org, Isabel Openshaw. I.Openshaw@kew.org
}
