% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stand_tariff}
\alias{stand_tariff}
\title{Tariff number by stand height}
\usage{
stand_tariff(spcode, height, sigma_h = NA)
}
\arguments{
\item{spcode}{species code}

\item{height}{tree height in metres}

\item{sigma_h}{sigma for height (optional)}
}
\value{
either tariff number or if sigma_h is provided, then returns a list
of the tariff number and uncertainty
}
\description{
Use the estimated stand top height to calculate the stand
tariff number.
}
\references{
Jenkins, Thomas AR, et al. "FC Woodland Carbon Code:
Carbon Assessment Protocol (v2. 0)." (2018).
}
\author{
Justin Moat. J.Moat@kew.org, Isabel Openshaw. I.Openshaw@kew.org
}
