% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WCC_functions.R
\name{tariff_vol_area}
\alias{tariff_vol_area}
\title{Tariff number from volume and basal area}
\usage{
tariff_vol_area(vol, dbh, sig_vol = NULL, re_dbh = 0.025, re = 0.025)
}
\arguments{
\item{vol}{tree volume (metres cubed)}

\item{dbh}{diameter at breast height (centimetres)}

\item{sig_vol}{sigma for tree volume (optional)}

\item{re_dbh}{relative measurement error for diameter at breast height (default is 0.025)}

\item{re}{relative error of coefficients (default = 0.025)}
}
\value{
Tariff number or if sigma for inputs are provided, then will return
a list of tariff number and sigma for tariff
}
\description{
Using the sample tree’s basal area and volume to calculate the
tariff number. Basal area is calculated by ba = (pi * dbh^2)/40000.
}
\examples{
tariff_vol_area(vol=0.5, dbh=24, sig_vol = 0.001, re_dbh = 0.05)
}
\references{
Jenkins, Thomas AR, et al. "FC Woodland Carbon Code:
Carbon Assessment Protocol (v2. 0)." (2018). (Equation 1)
}
\author{
Justin Moat. J.Moat@kew.org, Isabel Openshaw. I.Openshaw@kew.org
}
