% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tariff_vol_area}
\alias{tariff_vol_area}
\title{Tariff number from volume and basal area}
\usage{
tariff_vol_area(vol, dbh, sig_vol = NA, sig_dbh = NA, conf_a = 0.05)
}
\arguments{
\item{vol}{tree volume in metres cubed}

\item{dbh}{diameter at breast height in centimetres}

\item{sig_vol}{sigma for tree volume (optional)}

\item{sig_dbh}{sigma for diameter at breast height (optional)}

\item{conf_a}{1 - the percentage confidence interval for coefficients or
the relative error of coefficients (default = 5\%)}
}
\value{
Tariff number or if sigma for inputs are provided, then will return
a list of tariff number and sigma for tariff
}
\description{
Using the sample tree’s basal area and volume to calculate the
tariff number. Basal area is calculated by ba = (pi * dbh^2)/40000.
}
\examples{
tariff_vol_area(vol=0.5, dbh=24, sig_vol = 0.001, sig_dbh = 1)
}
\references{
Jenkins, Thomas AR, et al. "FC Woodland Carbon Code:
Carbon Assessment Protocol (v2. 0)." (2018). (Equation 1)
}
\author{
Justin Moat. J.Moat@kew.org, Isabel Openshaw. I.Openshaw@kew.org
}
