% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tariff_vol_area}
\alias{tariff_vol_area}
\title{Tariff number from volume and basal area}
\usage{
tariff_vol_area(vol, dbh, sigma_vol = NA, sigma_dbh = NA)
}
\arguments{
\item{vol}{tree volume in metres cubed}

\item{dbh}{diameter at breast height in centimetres}

\item{simga_vol}{sigma for tree volume (optional)}

\item{simga_dbh}{sigma for diameter at breast height (optional)}
}
\value{
Tariff number or if sigma for inputs are provided, then will return
a list of tariff number and sigma for tariff
}
\description{
Using the sample tree’s basal area and volume to calculate the
tariff number. Basal area is calculated by ba = (pi * dbh^2)/40000.
}
\references{
Jenkins, Thomas AR, et al. "FC Woodland Carbon Code:
Carbon Assessment Protocol (v2. 0)." (2018). (Equation 1)
}
\author{
Justin Moat. J.Moat@kew.org, Isabel Openshaw. I.Openshaw@kew.org
}
