% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tariffs}
\alias{tariffs}
\title{Tariff number by stand height}
\usage{
tariffs(spcode, height, dbh = NA, re_h = NA, re_dbh = NA, re_a = 0.025)
}
\arguments{
\item{spcode}{species code (short)}

\item{height}{tree height in metres}

\item{dbh}{diameter at breast height in centimetres}

\item{re_h}{relative error of height measurement (optional)}

\item{re_dbh}{relative error for diameter at breast height (optional)}

\item{re_a}{relative error of coefficients (default = 2.5\%)}
}
\value{
either tariff number or if re_h is provided, then returns a list
of the tariff number and uncertainty
}
\description{
Use the estimated stand top height to calculate the stand
tariff number.
}
\examples{
tariffs("OK", height= 10, dbh = 20)
tariffs("OK", 10, 20, re_h = 0.1, re_dbh = 0.05)
tariffs(c("OK","NS", "NS", "SP", "SP"), c(10,10,5,10,10), c(20,20,10,20,NA), re_h = 0.01, re_dbh = 0.05)
}
\references{
Jenkins, Thomas AR, et al. "FC Woodland Carbon Code:
Carbon Assessment Protocol (v2. 0)." (2018).
}
\author{
Isabel Openshaw. I.Openshaw@kew.org, Justin Moat. J.Moat@kew.org
}
