% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tariffs}
\alias{tariffs}
\title{Tariff number by stand height}
\usage{
tariffs(spcode, height, dbh = NA, sig_h = NA, sig_dbh = NA, conf_a = 0.05)
}
\arguments{
\item{spcode}{species code (short)}

\item{height}{tree height in metres}

\item{dbh}{diameter at breast height in centimetres}

\item{sig_h}{sigma for height (optional)}

\item{sig_dbh}{sigma for diameter at breast height (optional)}

\item{conf_a}{1 - the percentage confidence interval for coefficients or
the relative error of coefficients (default = 5\%)}
}
\value{
either tariff number or if sig_h is provided, then returns a list
of the tariff number and uncertainty
}
\description{
Use the estimated stand top height to calculate the stand
tariff number.
}
\examples{
tariffs("OK", height= 10, dbh = 20)
tariffs("OK", 10, 20, sig_h = 1, sig_dbh = 1)
tariffs(c("OK","NS", "NS", "SP", "SP"), c(10,10,5,10,10), c(20,20,10,20,NA), sig_h = 1, sig_dbh = 1)
}
\references{
Jenkins, Thomas AR, et al. "FC Woodland Carbon Code:
Carbon Assessment Protocol (v2. 0)." (2018).
}
\author{
Isabel Openshaw. I.Openshaw@kew.org, Justin Moat. J.Moat@kew.org
}
