% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{WD_Zanne}
\alias{WD_Zanne}
\title{Wood Density Dataset from Zanne et al.}
\format{
A data frame with 16,468 rows and 11 variables:
\describe{
\item{Family}{Character. Plant family.}
\item{Genus}{Character. Plant genus.}
\item{Species}{Character. Plant species.}
\item{Binomial}{Character. Genus + species name.}
\item{Wood.density}{Numeric. Wood density (g/cm³)}
\item{Reference}{Integer. Reference ID for source of data.}
\item{Continent}{Character. Continent where data was collected.}
\item{Biome}{Character. Biome classification.}
\item{Region}{Character. Combined continent and biome description.}
\item{wd_sd}{Numeric. The calculated standard deviation of the wood density
estimates from the available data for that species if there are more than
10 estimates per class.}
\item{wd_global_sd}{Numeric. Global standard deviation of wood density
across all species.}
}
}
\source{
Zanne, A.E., et al. (2009). Global wood density database. Dryad. http://hdl.handle.net/10255/dryad.235
}
\usage{
WD_Zanne
}
\description{
A dataset of wood density measurements compiled by Zanne et al. across various species, continents, and biomes.
}
\keyword{datasets}
