% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{woodbiomass}
\alias{woodbiomass}
\title{Forestry commission wood biomass}
\usage{
woodbiomass(treevol, nsg)
}
\arguments{
\item{treevol}{tree volume in metres cubed}

\item{nsg}{Nominal Specific Gravity}
}
\value{
biomass in oven dry tonnes
}
\description{
Multiply the mean total tree volume by the nominal specific
gravity to give the biomass, in oven dry tonnes.
}
\references{
Jenkins, Thomas AR, et al. "FC Woodland Carbon Code:
Carbon Assessment Protocol (v2. 0)." (2018).
}
\author{
Justin Moat. J.Moat@kew.org, Isabel Openshaw. I.Openshaw@kew.org
}
